add uploading token api call to add token modal
