+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/AppleBusinessManagerPage/_styles.scss
@@ -43,17 +43,6 @@
     }
   }
 
-  // NOTE: STYLES FOR THE ADM SETUP STEPS
-  // h4 {
-  //   margin-bottom: 0;
-  //   font-size: $x-small;
-  // }
-
-  // p {
-  //   font-size: $x-small;
-  //   margin: 0 0 $pad-large;
-  // }
-
   &__url-inputs-wrapper {
     display: flex;
     flex-direction: column;
@@ -63,32 +52,6 @@
 
   &__url-input {
     margin-bottom: 0;
-  }
-
-  &__permissions-list {
-    margin-top: $pad-large;
-    list-style: disc;
-    display: flex;
-    flex-direction: column;
-    gap: $pad-medium;
-  }
-
-  &__request-button {
-    display: flex;
-    gap: $pad-small;
-    align-items: center;
-    margin-top: $pad-small;
-
-    label {
-      display: flex;
-      gap: $pad-small;
-      cursor: pointer;
-    }
-  }
-
-  &__button-wrap {
-    display: flex;
-    gap: $pad-medium;
   }
 
   // TODO: remove these styles when the updated error component is used
